requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
python-multipart
#openai
agno
####
pypdf
lancedb
##Embedders:
torch
transformers
sentencepiece
tantivy
###
streamlit
